Refactor settings.tsx into separate section components
Split the monolithic dashboard settings page (1,613 lines) into smaller,
focused components for better maintainability:

- AvatarUploadSection.tsx (186 lines): Avatar upload logic and UI
- DonationSettingsForm.tsx (210 lines): Donation configuration form
- PlatformConnectionsPanel.tsx (184 lines): Streaming platform connections
- UserRolesManagement.tsx (644 lines): Role management UI (invitations,
  team members, role granting)

The main settings.tsx is now 472 lines (composition only), well under
the 500 line target.
