+++ frontend/src/components/settings/index.ts
@@ -1,0 +1,4 @@
+export { default as AvatarUploadSection } from "./AvatarUploadSection";
+export { default as DonationSettingsForm } from "./DonationSettingsForm";
+export { default as PlatformConnectionsPanel } from "./PlatformConnectionsPanel";
+export { default as UserRolesManagement } from "./UserRolesManagement";
